refactor(dashboard): rename finance.ts to finance-auditor.ts and tighten dashboard role boundaries
